23680961: clean architecture?
