+++ delivery/requests.go
@@ -1,4 +1,4 @@
-package main
+package delivery
 
 type SignupRequest struct {
 	Login     string `json:"login"`
