--- user.go
@@ -1,7 +1,0 @@
-package main
-
-type User struct {
-	Login    string
-	Email    string
-	Password string
-}
